Renamed old LevelGameplayController to LevelGameplayPresenter and created new LevelGameplayController to handle tasks like spawning and creating  inventory items.
